feat: 获取role列表带分页；解析struct tag

+++ common/models/req.go
@@ -7,8 +7,8 @@ type LoginRequest struct {
 
 type RoleListRequest struct {
 	SearchKey string `json:"search_key"`
-	Page      int64  `json:"page"`
-	PageSize  int64  `json:"page_size"`
+	Page      int    `json:"page"`
+	PageSize  int    `json:"page_size"`
 }
 
 type RoleAddRequest struct {
